connected new UI to backend
