--- app/lib/MyCard.dart
@@ -1,50 +1,0 @@
-import 'dart:io';
-import 'package:flutter/material.dart';
-
-class MyCard extends StatefulWidget {
-  final String imagePath;
-  final String text;
-
-  MyCard({required this.imagePath, required this.text});
-
-  @override
-  _MyCardState createState() => _MyCardState();
-}
-
-class _MyCardState extends State<MyCard> {
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-      shape: RoundedRectangleBorder(
-        borderRadius: BorderRadius.circular(15.0),
-      ),
-      elevation: 5,
-      child: Column(
-        children: <Widget>[
-          ClipRRect(
-            borderRadius: BorderRadius.only(
-              topLeft: Radius.circular(15),
-              topRight: Radius.circular(15),
-            ),
-            child: Image.asset(
-              widget.imagePath, // Use the provided image path
-              fit: BoxFit.cover,
-              width: double.infinity,
-              height: 150,
-            ),
-          ),
-          Container(
-            padding: const EdgeInsets.all(8.0),
-            child: Text(
-              widget.text,
-              style: TextStyle(
-                fontSize: 16,
-                color: Colors.black,
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
